Add files via upload
If you are getting married and want to increase the statistical odds of you having an I DO FOR ALWAYS, this course is for you. BONUS... if you are getting married in Florida, most (if not all) clerks of the courts recognize our completion certificate to offer you a discount on your marriage license and even the option to waive the 3-day wait (cooling off) period to get married.